Refactor single sub expr visitor to abstract class
